# change fft image parametrization to use only one op for the batch