Current Lucid FFT image parametrization creates a separate `irfft` op per image in a batch. If you create just one FFT op and send all images through it, it is much faster. Essentially each `irfft` op you put in the graph incurs a ~7ms CPU cost for a forward pass and 8ms CPU cost per backward pass. These per op costs **are paid in series on each sess.run**!

Seems to be related to a per-op recomputation of an execution plan that's done each time in a session.run. This behavior is described here: https://github.com/tensorflow/tensorflow/issues/6541#issuecomment-525828793

Results on a V100 for a neuron vis with batch of 32 and 128 steps:
https://colab.research.google.com/drive/1jfQP9E6M8piGaHxy75Tl23VeTx-QscL9#scrollTo=VFN7L_QNuwOI
time with current parametrization: 45.98s
time with updated parametrization: 12.89s

TF timelines that show the issue with a batch of 16 items:
**before (250 ms):**
![Screenshot 2019-10-01 11 53 46](https://user-images.githubusercontent.com/800877/66002536-c54e2a80-e458-11e9-9f12-4b7ec47fb67c.png)

**after (30 ms):**
![Screenshot 2019-10-01 12 24 57](https://user-images.githubusercontent.com/800877/66002601-e7e04380-e458-11e9-94bd-a079ca0f1e59.png)