@ludwigschubert 2019-10-01
@michaelpetrov this is awesome, thank you! IMHO can be merged. :-)